Track changes to navqueue positions as a document is edited
Fixes # 1480.

We piggyback on Scintilla's indicators facility to remember positions
in a way that is robust in the face of document changes. When adding
a position to the navqueue, we fill that position with a special, hidden
indicator, setting its value to a unique ID. Later, we can enumerate
all ranges of this indicator in the document to find the new position
with that ID.

I don't even think this is much of a hack: this indicator could reasonably
be shown with INDIC_POINT style.

We still remember the original position, for fallback and optimization.

To avoid accumulating too many ranges of this indicator, we limit
the entire navqueue to MAX_NAVQUEUE_LENGTH. Ideally it should be
a preference rather than a constant...

Because we now need to talk to Scintilla on more navqueue operations,
we cache pointers to GeanyDocument in the navqueue, so we don't have
to call document_find_by_filename every time.

Because navqueue_add_position now takes a GeanyDocument rather than just
its file_name, the check for non-NULL file_name has been moved inside it.